GN rule to build set of different projects and compile into bootfs
This adds a script that can invoke GN to generate ninja files to build
multiple projects and assemble them into a bootfs suitable for loading
into magenta. The list of projects is hardcoded (for now) and the
contents of each project are listed in the 'escher', 'fortune' and
'mojo' files.

Each project definition is a json object that must contain
a string "label" which is the GN label name of the root of that
project's build as well as a list of objects called "binaries" each
of which contains a string called "binary" that is the name of
the executable along with a string called "bootfs_path" that is the
location of that binary in the user bootfs.

To use, from the root of the fuchsia jiri checkout run

./packages/gen.py
ninja -C out/Debug

gen.py also supports -o / --outdir to output to a different path than
out/Debug, although it doesn't actually know anything about configurations.

This will result in a file called user.bootfs in the root of the outdir
that can be loaded by magenta as a userfs.

Change-Id: Iac07d6e230faee7ba740db794d9105a6aecafbe8