packages: Add Dart VM tests to the image
This CL adds two programs to the image. dart_vm_tests runs a single
Dart VM unittest that is named in a command line argument.
run_dart_vm_tests runs a list of tests contained in a file named
in a command line argument by invoking dart_vm_tests for each one.
The list of tests is also added to the image.

Change-Id: I95bbeffe3c8cedea0b0ade00f029d0cfe2f83d26