[gn mkbootfs] Remove depsfile from gn generator
The scheme of gen.py producing a depsfile in the out dir was too
fragile as 'ninja -t clean' or 'gn clean' or 'rm -rf out/.../gen' did
not understand that this file was important and the behavior with the
depsfile missing is ninja tries to run the mkbootfs step too early,
resulting in a hard-to-understand error. This gets rid of the depsfile
and instead requires that the configs spell a label that results in
the binary specified being built, which means the mkbootfs rule will
generally run last.

This still has the issue that the generated ninja files point to files
in //packages/gn/mkbootfs which is not in the outdir and so multiple
out dirs can get confused. This is tricky to resolve but not new. The
right answer is probably to put the configuration information into the
gn args and run the amalgamation logic at GN generation time instead
of before so all of the outputs can be preserved.

Address issue US-35

Change-Id: I40ff30c317d5d63a948c4fd567745f615a4adba4