Propragate process_modules.py failure.
Currently, if gn/process_modules.py fails to load a JSON file for a module
(e.g. because of a syntax error), it will print the error to stdout and
keep going.  Since stdout is passed to gn/BUILD.gn, this can make it
tricky to determine where the syntax error is.  This change prints the
error instead to stderr and fails the build by returning a nonzero exit
code.

For example, previously a missed comma resulted in:
  ERROR at //packages/gn/BUILD.gn:24:18: Unable to load "packages/gn/char 121/BUILD.gn".
    all_binaries = exec_script("process_modules.py",
                   ^--------------------------------

With this change the output from gen.py is now:
  ERROR at //packages/gn/BUILD.gn:24:18: Script returned non-zero exit code.
    all_binaries = exec_script("process_modules.py",
                   ^----------
  Current dir: out/debug-x86-64/
  Command: /usr/bin/env -- packages/gn/process_modules.py --manifest out/debug-x86-64/gen/packages/gn/user.bootfs.manifest --modules default --build-root out/debug-x86-64 --depfile out/debug-x86-64/gen/packages/gn/mkbootfs.d --arch x64
  Returned 1.
  stderr:

  Failed to parse config packages/gn/cmdutils, error Expecting , delimiter: line 8 column 13 (char 121)

It also makes the data returned via stdout more explicit, by
favoring the python3-friendly sys.stdout.write over the non-function
version of print.

Change-Id: Ibbdbdfb9744c4ef3c4f4c40ee0a83ad771fb11a9